+++ Assets/Scripts/Vanilla/Frameworks/Entities/VehicleInteraction.cs
@@ -5,7 +5,7 @@ namespace MVZ2.Vanilla.Entities
     public static class VehicleInteraction
     {
         public const int CRUSH = 0;
-        public const int SPIKES = 1;
+        public const int IGNORE = 1;
         public const int BLOCK = 2;
     }
 }